Updated GroovyScript 1.1.3 -> 1.2.5 Remove deprecated usage of 'classes' inside runConfig.json Moved common.groovy to account for these changes and subsequentially changed the import of the files who depended on it
