+++ manifest.json
@@ -660,7 +660,7 @@
     },
     {
       "projectID": 687577,
-      "fileID": 5635093,
+      "fileID": 6695118,
       "required": true
     },
     {
